+++ medusa/helpers/__init__.py
@@ -42,7 +42,7 @@ import guessit
 
 from medusa import app, db
 from medusa.common import DOWNLOADED, USER_AGENT
-from medusa.helper.common import (episode_num, http_code_description, media_extensions,
+from medusa.helper.common import (http_code_description, media_extensions,
                                   pretty_file_size, subtitle_extensions)
 from medusa.helpers.utils import generate
 from medusa.imdb import Imdb
@@ -739,37 +739,6 @@ def fix_set_group_id(child_path):
                 u'Failed to respect the set-group-ID bit on the parent'
                 u' directory for {path} (setting group ID {gid})',
                 {'path': child_path, 'gid': parent_gid})
-
-
-def get_absolute_number_from_season_and_episode(series_obj, season, episode):
-    """Find the absolute number for a show episode.
-
-    :param show: Show object
-    :param season: Season number
-    :param episode: Episode number
-    :return: The absolute number
-    """
-    absolute_number = None
-
-    if season and episode:
-        main_db_con = db.DBConnection()
-        sql = 'SELECT * FROM tv_episodes WHERE indexer = ? AND showid = ? AND season = ? AND episode = ?'
-        sql_results = main_db_con.select(sql, [series_obj.indexer, series_obj.series_id, season, episode])
-
-        if len(sql_results) == 1:
-            absolute_number = int(sql_results[0]['absolute_number'])
-            log.debug(
-                u'Found absolute number {absolute} for show {show} {ep}', {
-                    'absolute': absolute_number,
-                    'show': series_obj.name,
-                    'ep': episode_num(season, episode),
-                }
-            )
-        else:
-            log.debug(u'No entries for absolute number for show {show} {ep}',
-                      {'show': series_obj.name, 'ep': episode_num(season, episode)})
-
-    return absolute_number
 
 
 def get_all_episodes_from_absolute_number(show, absolute_numbers, indexer_id=None, indexer=None):