+++ medusa/server/api/v2/config.py
@@ -83,12 +83,12 @@ class ConfigHandler(BaseRequestHandler):
         'showDefaults.status': EnumField(app, 'STATUS_DEFAULT', (SKIPPED, WANTED, IGNORED), int),
         'showDefaults.statusAfter': EnumField(app, 'STATUS_DEFAULT_AFTER', (SKIPPED, WANTED, IGNORED), int),
         'showDefaults.quality': IntegerField(app, 'QUALITY_DEFAULT', validator=Quality.is_valid_combined_quality),
-        'showDefaults.subtitles': BooleanField(app, 'SUBTITLES_DEFAULT', validator=lambda v: app.USE_SUBTITLES,
-                                               converter=bool),
+        'showDefaults.subtitles': BooleanField(app, 'SUBTITLES_DEFAULT', converter=bool),
         'showDefaults.seasonFolders': BooleanField(app, 'SEASON_FOLDERS_DEFAULT', validator=season_folders_validator,
                                                    converter=bool),
         'showDefaults.anime': BooleanField(app, 'ANIME_DEFAULT', converter=bool),
         'showDefaults.scene': BooleanField(app, 'SCENE_DEFAULT', converter=bool),
+        'showDefaults.showLists': ListField(app, 'SHOWLISTS_DEFAULT'),
         'anonRedirect': StringField(app, 'ANON_REDIRECT'),
         'emby.enabled': BooleanField(app, 'USE_EMBY'),
 
@@ -480,7 +480,8 @@ class ConfigHandler(BaseRequestHandler):
         'anime.anidb.username': StringField(app, 'ANIDB_USERNAME'),
         'anime.anidb.password': StringField(app, 'ANIDB_PASSWORD'),
         'anime.anidb.useMylist': BooleanField(app, 'ANIDB_USE_MYLIST'),
-        'anime.autoAnimeToList': BooleanField(app, 'AUTO_ANIME_TO_LIST')
+        'anime.autoAnimeToList': BooleanField(app, 'AUTO_ANIME_TO_LIST'),
+        'anime.showlistDefaultAnime': ListField(app, 'SHOWLISTS_DEFAULT_ANIME')
     }
 
     def get(self, identifier, path_param=None):
@@ -608,6 +609,7 @@ class DataGenerator(object):
         section_data['showDefaults']['seasonFolders'] = bool(app.SEASON_FOLDERS_DEFAULT)
         section_data['showDefaults']['anime'] = bool(app.ANIME_DEFAULT)
         section_data['showDefaults']['scene'] = bool(app.SCENE_DEFAULT)
+        section_data['showDefaults']['showLists'] = list(app.SHOWLISTS_DEFAULT)
 
         section_data['logs'] = {}
         section_data['logs']['debug'] = bool(app.DEBUG)
@@ -1219,5 +1221,6 @@ class DataGenerator(object):
                 'password': app.ANIDB_PASSWORD,
                 'useMylist': bool(app.ANIDB_USE_MYLIST)
             },
-            'autoAnimeToList': bool(app.AUTO_ANIME_TO_LIST)
+            'autoAnimeToList': bool(app.AUTO_ANIME_TO_LIST),
+            'showlistDefaultAnime': app.SHOWLISTS_DEFAULT_ANIME
         }