+++ medusa/server/api/v2/series.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals
 
 import logging
 
-from medusa import ws
+from medusa import app, ws
 from medusa.logger.adapters.style import BraceAdapter
 from medusa.server.api.v2.base import (
     BaseRequestHandler,
@@ -15,7 +15,7 @@ from medusa.server.api.v2.base import (
     iter_nested_items,
     set_nested_value
 )
-from medusa.tv.series import Series, SeriesIdentifier
+from medusa.tv.series import SaveSeriesException, Series, SeriesIdentifier
 
 from six import itervalues, viewitems
 
@@ -93,15 +93,35 @@ class SeriesHandler(BaseRequestHandler):
         if not identifier:
             return self._bad_request('Invalid series identifier')
 
-        series = Series.find_by_identifier(identifier)
-        if series:
+        if Series.find_by_identifier(identifier):
             return self._conflict('Series already exist added')
 
-        series = Series.from_identifier(identifier)
-        if not Series.save_series(series):
-            return self._not_found('Series not found in the specified indexer')
+        data_options = data.get('options', {})
 
-        return self._created(series.to_json(), identifier=identifier.slug)
+        try:
+            options = {
+                'default_status': data_options.get('status'),
+                'quality': data_options.get('quality', {'preferred': [], 'allowed': []}),
+                'season_folders': data_options.get('seasonFolders'),
+                'lang': data_options.get('language'),
+                'subtitles': data_options.get('subtitles'),
+                'anime': data_options.get('anime'),
+                'scene': data_options.get('scene'),
+                'paused': data_options.get('paused'),
+                'blacklist': data_options['release'].get('blacklist', []) if data_options.get('release') else None,
+                'whitelist: ': data_options['release'].get('whitelist', []) if data_options.get('release') else None,
+                'default_status_after': data_options.get('statusAfter'),
+                'root_dir': data_options.get('rootDir'),
+                'show_lists': data_options.get('showLists')
+            }
+
+            queue_item_obj = app.show_queue_scheduler.action.addShow(
+                identifier.indexer.id, identifier.id, data_options.get('showDir'), **options
+            )
+        except SaveSeriesException as error:
+            return self._not_found(error)
+
+        return self._created(data=queue_item_obj.to_json)
 
     def patch(self, series_slug, path_param=None):
         """Patch series."""